// backend/src/shared/@enum/marital-status.enum.ts
export enum EMaritalStatus {
    Single = "single",
    Married = "married",
    Divorced = "divorced",
    Widowed = "widowed",
    Separated = "separated",
    DomesticPartnership = "domestic partnership",
    CivilUnion = "civil union",
    Engaged = "engaged",
    Cohabiting = "cohabiting",
    Annulled = "annulled",
    Minor = "minor"
  }