// backend/src/shared/@enum/realtionships.enum.ts
export enum ERelationships {
    Father = "father",
    Mother = "mother",
    Son = "son",
    Daughter = "daughter",
    Husband = "husband",
    Wife = "wife",
    Brother = "brother",
    Sister = "sister",
    Grandfather = "grandfather",
    Grandmother = "grandmother",
    Grandson = "grandson",
    Granddaughter = "granddaughter",
    Uncle = "uncle",
    Aunt = "aunt",
    Nephew = "nephew",
    Niece = "niece",
    Cousin = "cousin",
    Other = "other"
  }
  